+++ main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from app.routers import user_router, admin_router, moderator_router, article_router, favorite_router, auth_router, search_router, upload_router
+from app.routers import user_router, admin_router, moderator_router, article_router, favorite_router, auth_router, search_router, upload_router, moderation_router
 
 app = FastAPI()
 
@@ -26,5 +26,7 @@ app.include_router(moderator_router.router, prefix="/moderators", tags=["moderat
 app.include_router(article_router.router, prefix="/articles", tags=["articles"])
 app.include_router(search_router.router, prefix="/search", tags=["search"])
 app.include_router(favorite_router.router, prefix="/favorites", tags=["favorites"])
+
+app.include_router(moderation_router.router, prefix="/moderation", tags=["moderation"])
 
 app.include_router(upload_router.router, prefix="/upload", tags=["upload"])
